Change the ownership of system paths.
We changed the ownership of the system paths. Previouly, they are
managed by sub-command classes respectively. Now we changed this
ownership to base command class (gf_cmd_base) because the paths are
common with each sub-commands.

With this modification, several sub-commands doesn't work sufficiently.
We will fix them in few days.