The gf_site module has been re-designed.
Previously, We defined a gf_site module handle as a wrapper for the
XML document tree (xmlDocPtr) but it found that proper process of the
complecated site management cannot be done by that implementation. So,
we will re-define the new gf_site module. It extracts a site
information to this module in memory and read from or write to the XML
file. This XML file will be treated as an input file of the XSLT
processing.

In addition, related module gf_date and gf_string is newly added.