Add XML schema for meta.gf file.
A meta.gf file is a file that describes a content which is located there.
For now, you need to write this file by hand, though, you can use this
schema grayfish.rnc file.

A grayfish.rnc is a XML schema written in RELAX NG Compact Syntax. You can
use this schema when you editing meta.gf to make a valid file.